requirements.txt

Django==4.0.1
Pillow==9.3.0
json5==0.9.6
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
asgiref==3.11.0



